Move pkg netpolerrors one level up
