+++ roxctl/netpol/connectivity/map/cta.go
@@ -13,7 +13,7 @@ import (
 	"github.com/stackrox/rox/pkg/errox"
 	"github.com/stackrox/rox/roxctl/common/environment"
 	"github.com/stackrox/rox/roxctl/common/npg"
-	"github.com/stackrox/rox/roxctl/netpol/connectivity/netpolerrors"
+	"github.com/stackrox/rox/roxctl/netpol/netpolerrors"
 	"k8s.io/cli-runtime/pkg/resource"
 )
